+++ consensus/helpers.go
@@ -26,7 +26,6 @@ const (
 	ByzantineThreshold = float64(2) / float64(3)
 
 	HotstuffMessageContentType = "consensus.HotstuffMessage"
-	UtilityMessageContentType  = "consensus.UtilityMessage"
 )
 
 var (